feat(cartoes,financeiro): reconciliation depth, Central-style table, posto picker search
Cartões — conciliação
- Add "repasse sem venda" (mirror of "sem repasse"): acquirer paid a lote with no
  system sale → likely unlaunched venda; surfaced in Resultado + Detalhamento.
- Fix diaLiquidacao to key on `vencimento` (bom-para) instead of dataPagamento,
  so not-yet-due credit shows "Aguardando" instead of a false "sem repasse".
- Reframe "sem repasse" copy: "não localizado no EDI — confirme no portal", not
  a verdict of non-payment. "Vendedor não identificado" instead of "—".
- Detalhamento: expected net by contract (bruto − comissão% − fixed fee), portal
  lookup ficha (NSU/Aut) in copy + CSV, seller concentration alert, lote reference
  from /CARTAO_REMESSA, search + sort + bandeira filter, category cards.
- Resultado: split "Não conciliado" into A resolver / A receber (aguardando) /
  Conciliado; group by posto (collapsed by default) with subtotals; restyle to the
  Central BenchmarkSetor language (section headers, BarCell, "Maior diferença"
  badge, Total row).

Financeiro
- Single source lib/instrumentos: client debt = títulos pendentes (no longer drop
  convertido=true, which hid overdue receivables); tables + dashboard reconcile.
- Dashboard "Saldo em aberto": separate client cobrança from cartões/apps to
  compensate (kept honest, inadimplência not diluted); cartões flagged
  "não conciliado" with overdue-suspect warning.
- ReceberTabela: client cobrança only (cartões/apps removed → Todos/Atrasados
  matches the "Em atraso" card); título counts per Período pill.
- Remove Títulos em atraso block + DashboardMensal tab (renamed Inteligência →
  Dashboard, layout v6→v7); Apps split from Cartões via /ADMINISTRADORA.

CompanySelect: search box + explicit checkbox squares in the posto picker.

Docs: apurar-cron warns every redeploy needs --no-verify-jwt (401 gateway root
cause); apuracao_formas_pagamento venda codes → bigint (demo tenant overflow).

+++ docs/supabase-apuracao-caixas-pagamentos.sql
@@ -50,8 +50,10 @@ create index if not exists idx_apuracao_caixas_rede_empresa_data
 create table if not exists public.apuracao_formas_pagamento (
   rede_id uuid not null references public.redes(id) on delete cascade,
   empresa_codigo int not null,
-  venda_codigo int not null,
-  venda_prazo_codigo int not null,
+  -- bigint: alguns códigos de venda (ex.: tenant de demonstração sintético)
+  -- passam do teto de int4 (~2,1 bi). Reais cabem em int, mas bigint é seguro.
+  venda_codigo bigint not null,
+  venda_prazo_codigo bigint not null,
   data_movimento date,
   vencimento date,
   -- Forma
